% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template_app.R
\name{use_template_app}
\alias{use_template_app}
\title{Use a Shiny app template for the current package}
\usage{
use_template_app(path = "inst/shiny", overwrite = FALSE)
}
\arguments{
\item{path}{The path where the Shiny app will be created}

\item{overwrite}{set to \code{TRUE} if you want to force the Shiny app to
overwrite any existing Shiny apps in the given \code{path}}
}
\description{
Generate a skeleton Shiny app for the current package. This function must be
called from the root directory of an R package.
}

